+++ application/migrations/015_payment.php
@@ -12,7 +12,7 @@ class Migration_payment extends CI_Migration {
         `sender` varchar(255) NOT NULL,
         `recipient` varchar(255) NOT NULL,
         `transfer_time` date NOT NULL,
-        `amount` bigint(20) NOT NULL,
+        `amount` double NOT NULL,
         `urutan` INT(11) UNIQUE NOT NULL AUTO_INCREMENT,
         PRIMARY KEY (`uuid`)
       ) ENGINE=InnoDB DEFAULT CHARSET=utf8
